Update tsconfig to handle post-upgrade type issues
With the upgrade to vite 7.1.5 we also had to bump the vite plugin
for react to 5.0.2. However, when the plugin was updated it required
another config option for `moduleResolution`, which was not compatible
with our `module` config option.

The emitted code built by typescript is now a lot more modern than
previous builds.